no sid in url needed anymore, just define your iptv mux like this: http://localhost:17999/stream/service/9b3f60016e6199ea70e56dfa0c015890?profile=pass&descramble=0&emm=1
define your pass profile with rewrite PAT and rewrite SDT, don't rewrite PMT. The uuid is the service on satellite you want to descramble. TVH will find a single service on the mux which can be mapped to a channel.
This is currently not yet working and being investigated, for now you can test with the mux url in eg VLC directly.